Refactor embeddings and WhatsApp handler
